refatora ViewMosaic para usar masonry-wall no layout
